Add an explicit cast to WDText. Awake from nib requirements now met

+++ Inkpad-Core/Model/WDText.m
@@ -60,7 +60,7 @@ NSString *WDAlignmentKey = @"WDAlignmentKey";
     [coder encodeObject:text_ forKey:WDTextKey];
     [coder encodeObject:fontName_ forKey:WDFontNameKey];
     [coder encodeFloat:fontSize_ forKey:WDFontSizeKey];
-    [coder encodeInt32:alignment_ forKey:WDAlignmentKey];
+    [coder encodeInt32:(int32_t)alignment_ forKey:WDAlignmentKey];
 }
 
 - (id)initWithCoder:(NSCoder *)coder
